refactor: database와 service레이어로 분리하여 행위와 도메인을 분리
